refactor: move internal signing logic to signing pkg
